Fixes #2821 - Autoload paths
Also PowerManager children must be in namespace in order to autoload
